Sanity check remote merge

+++ bridge_monitor/scripts/import_block_meta_rsk.py
@@ -5,7 +5,6 @@ from datetime import datetime, timezone
 from tempfile import TemporaryFile
 import json
 import configparser
-from math import floor
 from typing import List
 
 from pyramid.paster import setup_logging
@@ -32,6 +31,9 @@ def parse_args(argv: List[str]):
         required=False,
     )
     parser.add_argument("--empty", action="store_true", help="Initialize empty")
+    parser.add_argument(
+        "--truncate", action="store_true", help="Truncate existing table"
+    )
 
     return parser.parse_args(argv[1:])
 
@@ -50,19 +52,20 @@ def get_blockchain_meta(dbsession: Session, name, expected_safe_limit=12):
 
 
 def write_from_file_to_db(
-    dbsession: Session, file
+    dbsession: Session, file, args
 ):  # file must be formatted as: [123123, 545343]\n[123124, 545344]\n...
     block_chain_meta = get_blockchain_meta(dbsession, "rsk")
-    dbsession.query(BlockInfo).filter(
-        BlockInfo.block_chain_id == block_chain_meta.id
-    ).delete()
+    if args.truncate:
+        dbsession.query(BlockInfo).filter(
+            BlockInfo.block_chain_id == block_chain_meta.id
+        ).delete()
 
     for count, line in enumerate(file):
         try:
             block_n, ts = json.loads(line.decode("utf-8"))
         except (
             ValueError
-        ):  # in a properly formatted an empty last line would trigger this
+        ):  # in a properly formatted file an empty last line would trigger this
             break
         dbsession.add(
             BlockInfo(
@@ -75,30 +78,19 @@ def write_from_file_to_db(
             dbsession.flush()
 
 
-def write_from_parquet_to_db(dbsession: Session, path: str):
+def write_from_parquet_to_db(dbsession: Session, path: str, args):
     block_chain_meta = get_blockchain_meta(dbsession, "rsk")
-    dbsession.query(BlockInfo).filter(
-        BlockInfo.block_chain_id == block_chain_meta.id
-    ).delete()
+    if args.truncate:
+        dbsession.query(BlockInfo).filter(
+            BlockInfo.block_chain_id == block_chain_meta.id
+        ).delete()
 
     df = pd.read_parquet(path)
     df.drop_duplicates(subset=["block_number"], inplace=True)
-    chunk_count = 20
-    chunk_length = floor(len(df) / chunk_count)
-    logger.info("Parquet length: %d, chunk count: %d", len(df), chunk_count)
-    for i in range(chunk_count):
-        start = i * chunk_length
-        end = (i + 1) * chunk_length if i + 1 != chunk_count else len(df) + 1
-        logger.info("Processing chunk %d blocks %d-%d", i, start, end)
-        for _, row in df[start:end].iterrows():
-            dbsession.add(
-                BlockInfo(
-                    block_chain_id=block_chain_meta.id,
-                    block_number=int(row["block_number"]),
-                    timestamp=datetime.fromtimestamp(row["timestamp"]),
-                )
-            )
-        dbsession.flush()
+    df.insert(0, "block_chain_id", block_chain_meta.id, allow_duplicates=True)
+    df["timestamp"] = pd.to_datetime(df["timestamp"], utc=False, unit="s")
+    df["timestamp"] = df["timestamp"].dt.tz_localize("UTC")
+    df.to_sql(BlockInfo.__tablename__, dbsession.bind, if_exists="append", index=False)
     del df  # df will likely be large, so del it just to be sure it's gone
